+++ pyproject.toml
@@ -28,6 +28,7 @@ dependencies = [
     "langdetect>=1.0.9",
     "pysbd>=0.3.4",
     "aiohttp>=3.12.13",
+    "tomli>=2.2.1",
 ]
 authors = [{ name = "MrXnneHang", email = "XnneHang@gmail.com" }]
 keywords = []
@@ -51,7 +52,7 @@ Issues = "https://github.com/XnneHangLab/XnneHangLab/issues"
 [project.scripts]
 test = "lab.test:main"
 acgo = "lab.__main__:main"
-get_root = "lab.utils.get_root:main" # 获取项目所在根目录绝对路径, 防止 streamlit UI 中默认以 `.` 作为根目录而找不到一些文件.
+get_root = "lab.config_manager.abs_root:main" # 获取项目所在根目录绝对路径, 防止 streamlit UI 中默认以 `.` 作为根目录而找不到一些文件.
 tmp = "lab.tmp:main" # debug 时随意写的一些脚本
 debug = "lab.debug:main"
 test-database="lab.database.test:main"
