+++ src/lab/api/core_logic.py
@@ -9,8 +9,8 @@ from typing import TYPE_CHECKING, Any
 from funasr import AutoModel  # 导入仍然在代码顶部，但只执行一次
 
 from lab._dataclass import RunnerSettings
+from lab.config_manager.config import load_settings_file
 from lab.models.lazy_model import generate_asr_results, generate_vad_results
-from lab.utils.config import load_settings_file
 from lab.utils.console.logger import Logger
 
 if TYPE_CHECKING:
