+++ src/lab/api/openai.py
@@ -10,10 +10,6 @@ from dotenv import load_dotenv
 
 from lab._session_keys import session_keys
 
-# from chatbot.config_manager import ServiceSettings, load_settings_file
-# from lab.utils.console.logger import Logger
-
-
 load_dotenv()
 
 sdk_base_url = "https://api.lingyiwanwu.com"
