+++ src/lab/config_manager/abs_root.py
@@ -8,11 +8,14 @@ from __future__ import annotations
 from pathlib import Path
 
 from lab._dataclass import RootAbsDir
-from lab.utils.config import load_settings_file, write_settings_file
+from lab.config_manager.config import load_settings_file, write_settings_file
+from lab.utils.console.logger import Logger
 
 
 def main():
     ROOT_DIR = Path(__file__).parent.parent.parent.parent
     settings = load_settings_file("root.toml", RootAbsDir)
     settings.root_dir = str(ROOT_DIR)
+    Logger.info(f"Set root directory to {settings.root_dir}")
+
     write_settings_file("root.toml", settings)
