+++ src/lab/test.py
@@ -10,8 +10,8 @@ import torchaudio
 from todo.__version__ import VERSION
 
 from lab._dataclass import RunnerSettings
+from lab.config_manager.config import load_settings_file
 from lab.models.lazy_model import FunASRModel, generate_asr_results, generate_sense_voice_results
-from lab.utils.config import load_settings_file
 from lab.utils.TxtHelper import split_into_words, split_into_words_no_punct
 
 if TYPE_CHECKING:
