+++ src/lab/ui.py
@@ -6,8 +6,8 @@ from pathlib import Path
 import streamlit as st
 
 from lab._dataclass import RootAbsDir
+from lab.config_manager.config import load_settings_file
 from lab.styles.global_style import style
-from lab.utils.config import load_settings_file
 
 style(True)
