+++ src/lab/utils/FFmpegHelper.py
@@ -4,7 +4,7 @@ import shutil
 from typing import TYPE_CHECKING
 
 from lab._dataclass import RunnerSettings
-from lab.utils.config import load_settings_file
+from lab.config_manager.config import load_settings_file
 from lab.utils.console.logger import Logger
 from lab.utils.SubprocessHelper import run_shell_command
